Add files via upload
Phyllosticta citricarpa. The pathogen P. citricarpa causes the disease citrus black spot. The published delimitation survey example (single epicenter) for this pest assumes two years since introduction, giving a 2,400 m square ‘potentially infested zone’ (EFSA, Lázaro, Parnell, Civera, Schans, Schenk, Schrader, et al. 2020). The outermost 800-m-wide band is inspected first. The area of that band is 7.7 km2, so the estimated number of trees (N) is 345,600. Only part of that area will be inspected, however, as the authors specified taking a hypergeometric sample size (n) of ~3,723 trees (C [confidence] = 0.95, p [infestation rate] = 0.001, sensitivity = 0.8; n varies slightly as N decreases). Based on the host density, the effective sample area is 0.083 km2. If no detections occur in Band 1, then the survey continues in an inner band (also 800 m in width), and so on, until perhaps reaching the innermost 1600-m square. Alternately, if pests are detected in Band 1, the survey expands outward, stopping only when no more pests are detected. The sample area and n are approximately the same in all inspected bands.

For this example, we simulated two scenarios that corresponded to the authors’ examples of narrowing down or enlarging the delimited area. In the first, the infested area was an 0.8-km-radius circle around the epicenter, smaller than the initial potentially infested area. In the second, the infested area radius was 3.2 km, larger than the potential infested area. The defined host density was 45,000 trees per km2 (EFSA, Lázaro, Parnell, Civera, Schans, Schenk, Schrader, et al. 2020).

DTDS survey design. We use the given tree density and infestation rate of concern to calculate m = 45 infected trees per km2. This gives Acrit = 1.33 km2. The pathogen is a weak disperser (Perryman, Clark, and West 2014). In both scenarios we matched R to the initial half-side-length of the potentially infested zone in the original plans (1.2 km in Scenario 1, and 2.4 km in Scenario 2). In both cases, based on R and m, the values of W0 were not practicable. As in the TBRFV example (Supp. Info. Fig. S3B), we added multiple transects, which equated to adding orchard rows, to achieve WT of about 5 m (i.e., 1 row) for one inspector. The values of NT were 108 in Scenario 1, and 55 in Scenario 2 (longer R), with an estimated row spacing of 4.7 m between trees, the effective DTDS sample size in both scenarios was about 56,170 trees (Table 1). Assuming 3 min per tree, the DTDS plans would take an additional 2,250 h to complete.