Add files via upload
Pine wood nematode (PWN). The delimiting survey plan for PWN (Bursaphelenchus xylophilus) in Sweden relies on both physical samples and traps (Schroeder 2012), but we focused on the physical sampling activity here. The plan states that complete sampling is not possible, due to the size of the area to be surveyed. The main sampling area radius is 8 km. The target sampling areas are clear cuts, which have an estimated total area of about 6 km2 (not shown). The plan does not specify exactly how many hosts to inspect, so we evaluated one version with exhaustive sampling (n ≈ 15,000), and one sampling version designed to give a 95 percent confidence in finding an infestation rate of 0.1, which has n = 300. Based on that value, a little more than two samples would be taken in every clear cut.

DTDS survey design. In this example, all clear cut areas within the survey area need to be visited, so the design is for transect specifications to be applied at all sites. The survey area has about 140 clear cuts on average, with a mean size of 0.2-km square, so we set R = 0.1 km. Based on report specifications (Schroeder 2012), m = 30 infested hosts per km2, which gave Acrit = 2.0 km2. Based on Ltot of 55.1 km (= 140 × 2 × 2 × 0.1), W0 = 4.6 m (0.0046 km), which doesn’t quite work for one inspector. Using three (parallel) transects gave WT = 3.0 m, and the needed n of 600+ hosts (Table 1).